settings: correctly handle channel masks for source selection

+++ app/src/main/java/io/github/tmarsteel/hqrecorder/recording/RecordingService.kt
@@ -289,7 +289,7 @@ class RecordingService : Service() {
                 return false
             }
 
-            val maxAvailableChannel = Channel(device.channelCounts.max().toUInt())
+            val maxAvailableChannel = Channel(device.channelCounts.max())
             val maxRequestedChannel = config.tracks.maxOfOrNull { it.leftOrMonoDeviceChannel.coerceAtLeast(it.rightDeviceChannel) }
             if (maxRequestedChannel == null || maxRequestedChannel > maxAvailableChannel) {
                 return false
